+++ CHANGELOG.md
@@ -5,6 +5,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ### [1.0.1] - 
+- [Feat #176](https://github.com/prb28/vscode-amiga-assembly/issues/176) Compress the extension with webpack
 - [Pull #173](https://github.com/prb28/vscode-amiga-assembly/pull/173) *dansalvato* Data Generator minor fixes
 
 ### [1.0.0] - 2021-11-27